Remove WorkerBase and introduce WorkerExporter
